requirements.txt
flask
requests
python-dotenv
lyricsgenius
flask-wtf
flask-sqlalchemy
flask-login
postgres